[Bug 2980] reduce number of warnings  - string formatting(arguments should be literals)  - applying constness where necessary  - removing bad consts that are superfluous  - avoid signed/unsigned clashes in conditionals (either by cast or type change)  - signed/unsigned and promotion conflicts  - add prototypes for function pointer tables  - force unsigned argument promotion in calls to 'ctype' functions (is{digit,cntrl,...})

+++ ntpd/ntp_io.c
@@ -3265,7 +3265,7 @@ read_refclock_packet(
 	/* TALOS-CAN-0064: avoid signed/unsigned clashes that can lead
 	 * to buffer overrun and memory corruption
 	 */
-	if (rp->datalen <= 0 || rp->datalen > sizeof(rb->recv_space))
+	if (rp->datalen <= 0 || (size_t)rp->datalen > sizeof(rb->recv_space))
 		read_count = sizeof(rb->recv_space);
 	else
 		read_count = (u_int)rp->datalen;
